搜图插件-实现SauceNAO搜图

+++ GreenOnions.NT.Plugins/GreenOnions.NT.HPictures/LimitCache.cs
@@ -4,9 +4,9 @@ namespace GreenOnions.NT.HPictures
 {
     public static class LimitCache
     {
-        public static ConcurrentDictionary<long, DateTime> GroupCoolDown { get; set; } = new ConcurrentDictionary<long, DateTime>();
-        public static ConcurrentDictionary<long, DateTime> WhiteGroupCoolDown { get; set; } = new ConcurrentDictionary<long, DateTime>();
-        public static ConcurrentDictionary<long, DateTime> PrivateMessageCoolDown { get; set; } = new ConcurrentDictionary<long, DateTime>();
-        public static ConcurrentDictionary<long, int> LimitNumber { get; set; } = new ConcurrentDictionary<long, int>();
+        public static ConcurrentDictionary<uint, DateTime> GroupCoolDown { get; set; } = new ConcurrentDictionary<uint, DateTime>();
+        public static ConcurrentDictionary<uint, DateTime> WhiteGroupCoolDown { get; set; } = new ConcurrentDictionary<uint, DateTime>();
+        public static ConcurrentDictionary<uint, DateTime> PrivateMessageCoolDown { get; set; } = new ConcurrentDictionary<uint, DateTime>();
+        public static ConcurrentDictionary<uint, int> LimitNumber { get; set; } = new ConcurrentDictionary<uint, int>();
     }
 }
